+++ custom_components/remote_devices/light.py
@@ -63,7 +63,7 @@ async def async_setup_entry(
             name=device_name,
             manufacturer="Airwit",
             model="Plafondventilator",
-            sw_version="0.10.0",
+            sw_version="0.11.0",
         )
 
     async_add_entities(